fix the print out of water level at waterPlant() by casting float to int

+++ week-04/day-2/src/garden_application/Garden.java
@@ -37,7 +37,7 @@ public class Garden {
     float needsWater = needsWaterCounter(waterAmount);
     float waterPortion = waterAmount / needsWater;
 
-    System.out.printf("Watering with %f\n", waterAmount);
+    System.out.printf("Watering with %d\n", (int)waterAmount);
 
     for (Flower currentFlower : flowers) {
       if (currentFlower.needsWater()) {
